+++ resources/lang/en/validation.php
@@ -124,84 +124,8 @@ return [
     |
     */
 
-    // TO DO : Make custom validation messages files which entries would be referenced in the messages method of request objects.
-
     'custom' => [
-        'borrowedItems' => [
-            'required' => 'Select games to borrow.'
-        ],
-        'borrowedItems.*' => [
-            'inventory_item_available' => 'The game :item is not available anymore.'
-        ],
-        'borrowerEmail' => [
-            'required' => 'Enter the borrower\'s registered email address.',
-            'email' => 'The email address\' format is incorrect.',
-            'exists' => 'This email address is not registered in our database.'
-        ],
-        'borrowerPassword' => [
-            'required' => 'Enter the borrower\'s password.',
-            'password_for' => 'The password is incorrect.'
-        ],
-        'startDate' => [
-            'required' => 'Choose the borrowing\'s start date.',
-            'after_or_equal' => 'The borrowing\'s start date must be superior or equal to the current date.'
-        ],
-        'expectedReturnDate' => [
-            'required' => 'Choose the borrowing\'s expected return date.',
-            'after_or_equal' => 'The borrowing\'s expected return date must be superior or equal to the current date.'
-        ],
-        'guarantee' => [
-            'required' => 'Fill in the guarantee.',
-            'numeric' => 'The guarantee must be a positive number.',
-            'regex' => 'The guarantee must be a positive number with at most 2 decimals.'
-        ],
-        'agreementCheck1' => [
-            'required' => 'You must agree to these terms to proceed.',
-            'accepted' => 'You must agree to these terms to proceed.'
-        ],
-        'agreementCheck2' => [
-            'required' => 'You must agree to these terms to proceed.',
-            'accepted' => 'You must agree to these terms to proceed.'
-        ],
-        'selectedBorrowings' => [
-            'required' => 'Select borrowings to end.'
-        ],
-        'durationMin' => [
-            'integer' => 'The minimal duration must be an integer.',
-            'min' => 'The minimal duration must be positive.'
-        ],
-        'durationMax' => [
-            'integer' => 'The maximal duration must be an integer.',
-            'min' => 'The maximal duration must be positive.',
-            'gte' => 'The maximal duration must be greater or equal to the minimal duration.'
-        ],
-        'playersMin' => [
-            'integer' => 'The minimal number of players must be an integer.',
-            'min' => 'The minimal number of players must be stricly positive.'
-        ],
-        'playersMax' => [
-            'integer' => 'The maximal number of players must be an integer.',
-            'min' => 'The maximal number of players must be stricly positive.',
-            'gte' => 'The maximal number of players must be greater or equal to the minimal number of players.'
-        ],
-        'genres' => [
-            'required' => 'Please fill in at least one genre for this item.'
-        ],
-        'genres.*' => [
-            'exists' => 'A selected genre doesn\'t exist in our database.'
-        ],
-        'nameFr' => [
-            'required' => 'Please fill in the item\' french name.'
-        ],
-        'nameEn' => [
-            'required' => 'Please fill in the item\' english name.'
-        ],
-        'inventoryItemId' => [
-            'required' => 'Please select an inventory item to delete.',
-            'integer' => 'Please select a valid inventory item.',
-            'exists' => 'The selected inventory item doesn\'nt exist in our database',
-            'inventory_item_not_borrowed' => 'A currently borrowed item can\'t be deleted.'
-        ]
+
     ],
 
     /*