+++ resources/lang/en/validation/deleteInventoryItem.php
@@ -1,0 +1,8 @@
+<?php
+
+return [
+    'inventoryItemId.required' => 'Please select an inventory item to delete.',
+    'inventoryItemId.integer' => 'Please select a valid inventory item by its id.',
+    'inventoryItemId.exists' => 'The selected inventory item doesn\'t exist in our database',
+    'inventoryItemId.inventory_item_not_borrowed' => 'A currently borrowed item can\'t be deleted.'
+];