+++ resources/lang/en/validation/endBorrowing.php
@@ -1,0 +1,5 @@
+<?php
+
+return [
+    'selectedBorrowings.required' => 'Select borrowings to end.'
+];