+++ resources/lang/en/validation/newBorrowing.php
@@ -1,0 +1,22 @@
+<?php
+
+return [
+    'borrowedItems.required' => 'Select games to borrow.',
+    'borrowedItems.*.inventory_item_available' => 'The game :item is not available anymore.',
+    'borrowerEmail.required' => 'Enter the borrower\'s registered email address.',
+    'borrowerEmail.email' => 'The email address\' format is incorrect.',
+    'borrowerEmail.exists' => 'This email address is not registered in our database.',
+    'borrowerPassword.required' => 'Enter the borrower\'s password.',
+    'borrowerPassword.password_for' => 'The password is incorrect.',
+    'startDate.required' => 'Choose the borrowing\'s start date.',
+    'startDate.after_or_equal' => 'The borrowing\'s start date must be superior or equal to the current date.',
+    'expectedReturnDate.required' => 'Choose the borrowing\'s expected return date.',
+    'expectedReturnDate.after_or_equal' => 'The borrowing\'s expected return date must be superior or equal to the current date.',
+    'guarantee.required' => 'Fill in the guarantee.',
+    'guarantee.numeric' => 'The guarantee must be a positive number.',
+    'guarantee.regex' => 'The guarantee must be a positive number with at most 2 decimals.',
+    'agreementCheck1.required' => 'You must agree to these terms to proceed.',
+    'agreementCheck1.accepted' => 'You must agree to these terms to proceed.',
+    'agreementCheck2.required' => 'You must agree to these terms to proceed.',
+    'agreementCheck2.accepted' => 'You must agree to these terms to proceed.'
+];