+++ resources/lang/fr/validation/endBorrowing.php
@@ -1,0 +1,5 @@
+<?php
+
+return [
+    'selectedBorrowings.required' => 'Sélectionnez des emprunts à terminer.'
+];